Change User Error Handling

+++ src/main/java/mvh/user/Female.java
@@ -42,7 +42,7 @@ public class Female extends User {
                 Age: %s\s
                 Weight in Kgs: %s\s
                 Height in centimeters: %s\s
-                """, userName, userGender, userAge, userWeight, userHeight);
+                """, userName, userGender, (int)userAge,(int) userWeight, (int)userHeight);
     }
 
 }
