+++ src/main/java/mvh/user/User.java
@@ -75,7 +75,7 @@ public class User {
                 Age: %s\s
                 Weight in Kgs: %s\s
                 Height in centimeters: %s\s
-                """, userName, userGender, userAge, userWeight, userHeight);
+                """, userName, userGender, (int)userAge, (int)userWeight, (int)userHeight);
     }
 
 }
